Refactored bootstrapper
- Renamed bootstrapper to testkit
- Moved CLI code out of main into cmd subpackage
